pyproject.toml
[project]
description = "Land Registry"
name = "land_registry"
version = "0.1.4"
requires-python = '>=3.11'
dependencies = [
    # Core dependencies for Cloud Functions
    "functions-framework>=3.0.0",
    "mangum>=0.19.0",
    # FastAPI and web framework
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "jinja2>=3.0.0",
    # Geospatial and data processing
    "folium>=0.14.0",
    "geopandas>=0.13.0",
    "pandas>=2.0.0",
    "shapely>=2.0.0",
    # Basic utilities
    "requests>=2.25.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    # AWS S3 integration (for catasto-2025 bucket - cadastral data)
    "boto3>=1.34.0",
    "pydantic-settings>=2.0.0",
    # Google Cloud Storage (for aecs4u-storage bucket - app/user data)
    "google-cloud-storage>=2.14.0",
    # Neon PostgreSQL database
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    # Panel for interactive data apps
    "panel>=1.3.0",
    "param>=2.0.0",
    "rich>=14.1.0",
    # AECS4U Authentication (includes Clerk integration) - from Google Artifact Registry
    "aecs4u-auth>=1.0.0",
    # AECS4U Storage abstraction (S3, GCS, Azure, Local) - from Google Artifact Registry
    "aecs4u-storage>=0.1.0",
    "moto>=5.1.13",
]

[tool.black]
line_length = 120
# target_version = ['py37', 'py38', 'py39', 'py310']
exclude = '__pycache__'

[tool.isort]
profile = "black"
filter_files = ["__pycache__"]

# Optional dependencies (for future use)
# [project.optional-dependencies]
# storage = ["aecs4u-storage>=0.1.0"]
# analytics = ["aecs4u-analytics>=0.1.0"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "coverage>=7.6.0",
    "moto[s3]>=4.2.0",
    "httpx>=0.25.0"
]

test = [
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0", 
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "coverage>=7.6.0",
    "tox>=4.0.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# Ignore archived/outdated tests
norecursedirs = ["tests/archive"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--cov=land_registry",
    "--cov-report=html:htmlcov",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-fail-under=20",
    "--ignore=tests/archive"
]
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests", 
    "slow: marks tests as slow",
    "parser: marks tests for parser functionality",
    "database: marks tests for database operations",
    "workflow: marks tests for Prefect workflows",
    "web: marks tests for web interface"
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
]

[tool.coverage.run]
source = ["land_registry"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod"
]

[tool.setuptools.packages.find]
where = ["land_registry"]

# =============================================================================
# UV Configuration - Private Package Sources
# =============================================================================

# Use keyring for authentication to Google Artifact Registry
[tool.uv]
keyring-provider = "subprocess"

# Define the Google Artifact Registry index for aecs4u packages
[[tool.uv.index]]
name = "aecs4u-gar"
url = "https://europe-west1-python.pkg.dev/apps-aecs4u/python-packages/simple/"
explicit = true

[tool.uv.sources]
# AECS4U packages from Google Artifact Registry
aecs4u-auth = { index = "aecs4u-gar" }
aecs4u-storage = { index = "aecs4u-gar" }
